Prune events past a given theshold
By default, the threshold is 500 events (maximumQueuedEventCount)

Events are now fetched ordered by date – this makes it so that if we need to delete any, we’re only deleting the oldest events from the list.

Additionally, the event list is pruned when attempting to send the events to the server.

Fix up project file