Rework the crash logging architecture
The new architecture makes it so that a developer can’t write data to `CrashLogging` directly.

Instead, they provide it with data via a `CrashLoggingDataProvider`.

The exception to this is the ability to request that `CrashLogging` update its knowledge of the current user. This can be called by the application in response to user-changing events (such as logging in, or changing email address).