Use for of Sodium to work around spec validation issue in Xcode 12
Because of the build system changes in Xcode 12 due to the support of
Apple Silicon, we now need to explicitly exclude the arm64 architecture
from the build, otherwise xcodebuild will fail to build.

I opened an issue in the Sodium repo, see
https://github.com/jedisct1/swift-sodium/issues/213, but I haven't
received feedback yet.

In order to unblock our progress, I created a fork of Sodium,
Sodium-Fork, which incorporate the necessary changes to pass the
validation.

I published the fork on CocoaPods, but unfortunately CI seems to fail to
find it.
See this build:
https://app.circleci.com/pipelines/github/Automattic/Automattic-Tracks-iOS/380/workflows/c0206d34-0f8a-4913-8b5d-b05e6dcb7c5b/jobs/936

I tried to set the `update-specs-sources` parameter to true, but it
didn't make any difference.
Interestingly, CI shows that the command runs, but there is no output
and the execution time is 1s, too little to fetch the CocoaPods specs
repo.
This makes me think there might be something wrong in the setup or in
how CI runs it.

To address this issue, I published the pod on my private specs repo, and
added its URL to the sources for the validation.

Funny enough, while the spec validation fails to find the new pod, the `pod
install` command in a different job has no problems with it.