Remove OCMock as a SPM dependency, add it as an XCFramework.
This is OCMock 3.8.1, downloaded from ocmock.org and built into an xcframework using the following command:

```bash
xcodebuild -create-xcframework
  -framework OCMock/iOS/OCMock.framework
  -framework OCMock/Mac/OCMock.framework
  -output OCMock.xcframework
```

We were previously pulling in OCMock using Cocoapods, but we are migrating from that to Swift Package Manager. However, OCMock uses unsafe flags in its SPM package manifest, and Xcode 12.5 does not allow dependencies on a package with unsafe flags in its manifest. (Xcode 13 does, so long as you pin to a particular commit, but as of this writing Xcode 13 is still in beta.)