Don't build tests unless we're doing the test action.
Because tests don't run on iOS devices, our OCMock xcframework doesn't include a slice for physical devices. That means that the tests fail to compile if you're building them for an iOS device destination. But we don't need that anyway, because Xcode already prevents the tests from running in that case.